moving base classes to __init__
